+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 ## Added
 - `SlurmScheduler` to perform distributed evaluation with slurm
 - `LocalScheduler` to perform locally evaluations using joblib
+- `Job` objects running on a `Scheduler` returning `Result` objects
 
+## Changed
+- `Base`-prefix is removed from all base classes which reside 
+ in `base.py` modules.
+ 
 ## [0.1.0] - 2019-07-27
 ### Added
 - `TensorboardReporter` result logger using `HParams`